requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
django-bootstrap3==14.2.0
django-filter==2.4.0
django-tables2==2.3.4
django==3.1.7
importlib-metadata==1.7.0; python_version < '3.8'
psycopg2-binary==2.8.4
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
zipp==3.4.0; python_version >= '3.6'
